% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_cities}
\alias{list_cities}
\title{List Available Cities for Finturk
Print available cities for Finturk quarterly data with plaka (license plate) numbers.}
\usage{
list_cities()
}
\value{
Data frame of available cities
}
\description{
List Available Cities for Finturk
Print available cities for Finturk quarterly data with plaka (license plate) numbers.
}
\examples{
list_cities()
}
