% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_groups}
\alias{list_groups}
\title{List available groups
Print available banking groups for a data source.}
\usage{
list_groups(source = c("bddk", "finturk"), lang = c("en", "tr"))
}
\arguments{
\item{source}{Either "bddk" or "finturk"}

\item{lang}{Either "tr" or "en" for names. "en" is default}
}
\value{
Data frame of available groups (invisibly)
}
\description{
List available groups
Print available banking groups for a data source.
}
\examples{
list_groups("bddk")
list_groups("finturk","tr")
}
