% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_tables}
\alias{list_tables}
\title{List Available Tables
Print available tables for a data source.}
\usage{
list_tables(source = c("bddk", "finturk"), lang = c("en", "tr"))
}
\arguments{
\item{source}{Either "bddk" or "finturk"}

\item{lang}{Either "tr" or "en" for column names. "en" is default}
}
\value{
Data frame of available tables (invisibly)
}
\description{
List Available Tables
Print available tables for a data source.
}
\examples{
list_tables("bddk")
list_tables("finturk", "tr")
}
